+++ redacted-compiler-plugin/src/main/kotlin/dev/zacsweers/redacted/compiler/RedactedCompilerPluginRegistrar.kt
@@ -2,12 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.redacted.compiler
 
+import dev.zacsweers.metro.compiler.compat.CompatContext
 import dev.zacsweers.redacted.BuildConfig.KOTLIN_PLUGIN_ID
 import dev.zacsweers.redacted.compiler.fir.RedactedFirExtensionRegistrar
-import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
 import org.jetbrains.kotlin.config.CompilerConfiguration
-import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrarAdapter
 import org.jetbrains.kotlin.name.ClassId
 
 public class RedactedCompilerPluginRegistrar : CompilerPluginRegistrar() {
@@ -18,6 +17,17 @@ public class RedactedCompilerPluginRegistrar : CompilerPluginRegistrar() {
 
   override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
     if (configuration[KEY_ENABLED] == false) return
+
+    val compatContext =
+      try {
+        CompatContext.create()
+      } catch (t: Throwable) {
+        System.err.println(
+          "[Redacted] Skipping enabling Redacted extensions, unable to create CompatContext"
+        )
+        t.printStackTrace()
+        return
+      }
 
     val replacementString = checkNotNull(configuration[KEY_REPLACEMENT_STRING])
     val redactedAnnotations =
@@ -33,11 +43,18 @@ public class RedactedCompilerPluginRegistrar : CompilerPluginRegistrar() {
         ClassId.fromString(it)
       }
 
-    FirExtensionRegistrarAdapter.registerExtension(
-      RedactedFirExtensionRegistrar(redactedAnnotations, unRedactedAnnotations)
-    )
-    IrGenerationExtension.registerExtension(
-      RedactedIrGenerationExtension(replacementString, redactedAnnotations, unRedactedAnnotations)
-    )
+    with(compatContext) {
+      registerFirExtensionCompat(
+        RedactedFirExtensionRegistrar(redactedAnnotations, unRedactedAnnotations)
+      )
+      registerIrExtensionCompat(
+        RedactedIrGenerationExtension(
+          replacementString,
+          redactedAnnotations,
+          unRedactedAnnotations,
+          compatContext,
+        )
+      )
+    }
   }
 }