+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 **Unreleased**
 --------------
 
+- Integrate Metro's compiler-compat infra to support a wider range of Kotlin compiler versions.
+- Test Kotlin `2.4.10-RC` and `2.4.20-Beta1`.
+
 1.17.0
 ------